+++ ici/include/memmgr.h
@@ -65,7 +65,7 @@ MemPtoAConverter memmgr_PtoA(   int mgrId);
 		/*	Returns the "PtoA" function of this memory
 			manager.					*/
 
-int		memmgr_open(int memKey, long memSize, char **memPtr,
+int		memmgr_open(int memKey, size_t memSize, char **memPtr,
 			uaddr *smId, char *partitionName,
 			PsmPartition *partition, int *memMgr,
 			MemAllocator afn, MemDeallocator ffn,
